Add existing file

+++ T6Model.h
@@ -40,8 +40,8 @@
 
 extern int stiffnessouter;
 extern int stiffnessinner;
-extern int pretensionouter;
-extern int pretensioninner;
+extern double pretensionouter;
+extern double pretensioninner;
 extern int dampinginner;
 extern int dampingouter;
